+++ README.md
@@ -6,11 +6,79 @@
 virtualenv -p python3 venv
     (or python3 -m venv venv)
 pip3 install -r requirements.txt
+```
+
+## Running as a framework
+_First of all: provide some arguments in the `main.py` file to collect information based on your data (WIP now, will be improved later)_  
+
+To run the framework:
+```bash
 python3 main.py
+```
+To run the tests:
+```bash
+chmod +x run_tests.sh
+./run_tests.sh
+```
+_or you can run them like this, of course:_
+```bash
+python3 -m unittest discover -v
+```
+## Running as a separated module
+Basic:
+```python3
+python3 -m src.scripts.<category>.<name> any_arguments_here
+```
+Example command:
+```bash
+python3 -m src.scripts.other.user_greeting JohnDoe
+```
+Example output:
+```
+{'message': "Successfully finished! (args: (), kwargs: {'username': "
+            "'johndoe'})",
+ 'result': 'Hello, JohnDoe!',
+ 'status': 'success'}
+
 ```
 
 ## Create your own script
 Use the following structure:  
+1. Create your own module directory in the following way:
+```
+/src/scripts/<choose_your_category_here>/<your_script_name>/<script_files>
+```
+2. Provide the following structure of your script directory:
+```
+your_script_name
+├── __init__.py     - use this module to set the default parent directory (you can copy this file from any other script)
+├── __main__.py     - use this module to provide some basic interface to use your script as a module (the same as if __name__ == "__main__")
+├── module.py       - use this module to describe the basic logic of your module (you can import it in the __main__.py to provide interface)
+└── test_module.py  - use this module for unittest tests
+```
+3. Create the `__init__.py` file. An example of the `__init__.py` boilerplate structure can be seen below:
+```python3
+import sys
+from pathlib import Path
+
+__root_dir = Path(__file__).parents[4]
+sys.path.append(str(__root_dir))
+
+```
+4. Create the `__main__.py` file. An example of the `__main__.py` boilerplate structure can be seen below:
+```python3
+#!/usr/bin/env python3
+
+from pprint import pprint
+from sys import argv
+
+from src.core.utils.module import run_module
+from .module import Runner
+
+result = run_module(Runner, args=argv, arg_name="username", arg_default="johndoe")
+pprint(result)
+```
+5. Create the module itself. An example of the basic `module.py` file can be seen below:
 ```python3
 #!/usr/bin/env python3
 
@@ -60,3 +128,4 @@ class Runner(OsintRunner):
         argument = kwargs.get("my_argument", "Arguments were not provided!")
         return ScriptResponse.success(message=f"Script finished with argument {argument}")
 ```
+6. For `test_module.py` you can use any required tests (as you wish). A test case for your module is required to keep the project clean.
