+++ requirements.txt
@@ -1,5 +1,18 @@
+aiodns==2.0.0
+aiohttp==3.6.2
+aiosmtpd==1.2
+async-timeout==3.0.1
+atpublic==1.0
+attrs==19.3.0
 certifi==2020.6.20
+cffi==1.14.0
 chardet==3.0.4
 idna==2.10
+mmh3==2.5.1
+multidict==4.7.6
+pycares==3.1.1
+pycparser==2.20
 requests==2.24.0
 urllib3==1.25.9
+verify-email==2.4.1
+yarl==1.4.2
