+++ run_tests.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+python3 -m unittest discover -v
