+++ src/scripts/other/simple_example/__init__.py
@@ -1,0 +1,5 @@
+import sys
+from pathlib import Path
+
+__root_dir = Path(__file__).parents[4]
+sys.path.append(str(__root_dir))
